+++ src/s2protocol.NET/Models/GameEvents/SBankFileEvent.cs
@@ -46,7 +46,7 @@ public record SUserOptionsEvent : GameEvent
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
 #pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
     {
-        
+
     }
