+++ src/s2protocol.NET/ReplayDecoder.cs
@@ -4,7 +4,6 @@ using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Scripting.Hosting;
 using s2protocol.NET.Models;
 using s2protocol.NET.Parser;
-using System.Diagnostics;
 using System.Globalization;
 using System.Runtime.CompilerServices;
 using System.Text;
